フラクタルサンプルプログラムの追加
